eagletest/requirement.txt

jdcal==1.4
Jinja2==2.10
kiwisolver==1.0.1
llvmlite==0.27.0
lxml==4.2.5
MarkupSafe==1.0
matplotlib==2.2.3
numba==0.42.0
numpy==1.14.5
openpyxl==2.5.12
ordereddict==1.1
packaging==17.1
pandas==0.23.4
pathlib==1.0.1
pathlib2==2.3.2
pickleshare==0.7.4
pip==9.0.3
prompt-toolkit==1.0.15
pyaes==1.6.1
pycparser==2.19
Pygments==2.2.0
pyOpenSSL==19.1.0
pyparsing==2.2.0
pyserial==3.4
python-dateutil==2.7.3
python-docx==0.8.7
python-version==0.0.2
pytz==2018.5
PyVISA==1.5
pywin32==222
